Lift and shift formatting of Pokemon data into new file
